+++ src/lib/Scenes/Map/MapContainer.tsx
@@ -4,13 +4,18 @@ import { MapRenderer } from "./MapRenderer"
 
 interface Props {
   citySlug?: string
+  initialCoordinates?: { lat: number; lng: number }
   hideMapButtons: boolean
 }
 
+/// This container is pretty simple, but it helps to have a simple component for the root of our ARMapContainerViewController.
 export class MapContainer extends Component<Props> {
   render() {
-    // TODO: If there is no city, we should show something else instead.
-    const { citySlug, hideMapButtons } = this.props
-    return citySlug ? <MapRenderer citySlug={citySlug} hideMapButtons={hideMapButtons} /> : <View />
+    const { citySlug, hideMapButtons, initialCoordinates } = this.props
+    return citySlug ? (
+      <MapRenderer citySlug={citySlug} hideMapButtons={hideMapButtons} initialCoordinates={initialCoordinates} />
+    ) : (
+      <View />
+    )
   }
 }